Støtt capitalize for dobbeltnavn med '-' for OMS
