+++ apps/omstillingsstoenad-ui/src/hooks/useInnloggetBruker.test.js
@@ -1,5 +1,5 @@
 import { renderHook } from '@testing-library/react'
-import useInnloggetBruker, { capitalize } from './useInnloggetBruker'
+import useInnloggetBruker from './useInnloggetBruker'
 
 const mock = jest.fn(async () => {
     return 'Ok'
@@ -28,35 +28,5 @@ describe('useInnloggetBruker', () => {
 
         expect(result.current).toBe(true)
         expect(mock).toBeCalledTimes(1)
-    })
-})
-
-describe('Test av capitalize', () => {
-    it('Tester at enkelt navn i uppercase får stor forbokstav', () => {
-        const testString = 'TEST'
-        const fasitString = 'Test'
-
-        expect(capitalize(testString)).toStrictEqual(fasitString)
-    })
-
-    it('Tester at dobbelt navn i uppercase får stor forbokstav', () => {
-        const testString = 'TEST TESTESEN'
-        const fasitString = 'Test Testesen'
-
-        expect(capitalize(testString)).toStrictEqual(fasitString)
-    })
-
-    it('Tester at enkelt navn i lowercase får stor forbokstav', () => {
-        const testString = 'test'
-        const fasitString = 'Test'
-
-        expect(capitalize(testString)).toStrictEqual(fasitString)
-    })
-
-    it('Tester at dobbelt navn i lowercase får stor forbokstav', () => {
-        const testString = 'test testesen'
-        const fasitString = 'Test Testesen'
-
-        expect(capitalize(testString)).toStrictEqual(fasitString)
     })
 })
